feat(dataclasses): Add dataclass style `__str__` to `c_class`
