+++ include/mlc/core/str.h
@@ -3,6 +3,7 @@
 #include "./object.h"
 #include <cctype>
 #include <cstring>
+#include <iomanip>
 #include <ostream>
 #include <sstream>
 #include <string_view>
