+++ python/mlc/dataclasses/utils.py
@@ -456,3 +456,9 @@ def prototype(
 
 def replace(obj: Any, /, **changes: Any) -> Any:
     return obj.__replace__(**changes)
+
+
+def stringify(obj: Any) -> str:
+    from mlc.core.func import Func
+
+    return Func.get("mlc.core.Stringify")(obj)
